backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
requests==2.31.0
beautifulsoup4==4.12.3
feedparser==6.0.11
trafilatura==1.6.0
groq==0.3.0
python-dotenv==1.0.1
pydantic==2.6.0
lxml==5.1.0
lxml_html_clean==0.1.0